pip-requirements.txt
pytz
Django==4.1
mysqlclient==2.1.1
django-crispy-forms==1.14.0
crispy-bootstrap5==0.6
tablib==3.2.1
django-filter==22.1
django-tables2==2.4.1
django-bootstrap-v5==1.0.11
django-import-export==2.9.0